Finished CMaze rework
CMaze interface is now ready to be used for multiple map generation types, however
it is still rather rudimentary and might not compile (I haven't tested it fully).

Also, some functions that don't a lot have been _inlined_ so as to speed them
up, as the compiler will now actually insert them into the calling function instead
of placing a direct to call to the function.