Working interface for Mazes
Mazes now actually have a virtual interface they inherit off of. Maze has become
an "abstract class" (not really since it has function implementations in it). However,
it is impossible to actually instantiate the CMaze class, because the constructors and
destructor (which is virtual, anyways) are protected, leaving the burder of class
construction up the derived class.

- Variables passed to constructor are now unsigned ints to prevent overflow/negative array values
- Constructors are now explicit
- Example maze class (GrowingTree) to demonstrate c++ style inheritance.